add test

+++ dolphinscheduler-remote/src/main/java/org/apache/dolphinscheduler/remote/utils/Constants.java
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.dolphinscheduler.remote.utils;
 
 import java.nio.charset.Charset;
@@ -21,7 +22,7 @@ import java.nio.charset.StandardCharsets;
 
 
 /**
- *  constant
+ * constant
  */
 public class Constants {
 
@@ -30,12 +31,12 @@ public class Constants {
     public static final String SLASH = "/";
 
     /**
-     *  charset
+     * charset
      */
     public static final Charset UTF8 = StandardCharsets.UTF_8;
 
     /**
-     *  cpus
+     * cpus
      */
     public static final int CPUS = Runtime.getRuntime().availableProcessors();
 
@@ -55,6 +56,6 @@ public class Constants {
     /**
      * warm up time
      */
-    public static final int WARM_UP_TIME = 10 * 60 * 1000;
+    public static final double WARM_UP_TIME = 10 * 60 * 1000;
 
 }
